+++ src/components/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod bottom_bar;
+pub(crate) mod central_panel;
+pub(crate) mod top_bar;
