+++ src/structs.rs
@@ -1,36 +1,3 @@
-use eframe::egui::{Rect, Response, Style};
-use rusqlite::Connection;
-use std::path::PathBuf;
-
-#[derive(Debug)]
-pub struct Invoicy {
-    pub max_file_size: usize,
-    pub image_file_path: Option<PathBuf>,
-    pub company_error_contact: Option<String>,
-    pub company_error_customer: Option<String>,
-    pub initialized: bool,
-    pub connection: Connection,
-    pub progress: f32,
-    pub style: Style,
-    pub file_name: String,
-    pub customer_selected: usize,
-    pub contact_selected: usize,
-    pub row_count: usize,
-    pub last_updated_row: usize,
-    pub table_data: Vec<(String, (usize, i32), (Rect, Response))>,
-    pub database_data_vec: Vec<DatabaseData>,
-    pub new_database_data_vec: Vec<DatabaseData>,
-    pub contact: Contact,
-    pub contacts: Vec<Contact>,
-    pub contact_form: bool,
-    pub customer: Customer,
-    pub customers: Vec<Customer>,
-    pub customer_form: bool,
-    pub current_row_value: DatabaseData,
-    pub totals: Vec<Total>,
-    pub grand_total: f64,
-}
-
 #[derive(Clone, Debug, PartialEq)]
 pub struct Total {
     pub value: f64,
